requirements.txt
fastapi
uvicorn
praat-parselmouth
numpy
noisereduce
moviepy
torch
python-multipart
azure-storage-blob
jwt
orjson
bson
pymongo